+++ Source/Falcon/src/Falcon.Web.Models/Api/Quest/SQuestDetail.cs
@@ -4,7 +4,7 @@
     {
         public int QuestNumber { get; set; }
         public string QuestTitle { get; set; }
-        public int PeopleScore { get; set; }
-        public int UserScore { get; set; }
+        public float PeopleScore { get; set; }
+        public float UserScore { get; set; }
     }
 }
